+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # 更新日志
 
+## [1.0.32]
+
+### 添加
+- /class/Timer.js
+- /class/Chatbox.js
+- /class/VideoPlayer.js
+
+### 修复
+- 无
+
+### 改进
+- 无
+
+### 更改
+- 无
+
+### 移除
+- 无
+
 ## [1.0.31]
 
 ### 添加
@@ -20,7 +39,7 @@
 ## [1.0.2]
 
 ### 添加
-- Dragger.js
+- /class/Dragger.js
 
 ### 修复
 - 无